Move v2 HCS runtime to internal package
