+++ internal/runtime/hcsv2/device.go
@@ -1,4 +1,6 @@
-package gcs
+// +build linux
+
+package hcsv2
 
 // TODO: JTERRY75 - When we fix the vendor problem this is from runc device.go.
 // Vendor it instead.
